converted rectangle coordinates to double, and made last edits to gameView

+++ src/GameView.java
@@ -19,7 +19,7 @@ public class GameView extends JComponent {
 		g.drawString("Hello World", 20, 20);
 		
 		g.setColor(Color.GRAY);
-		g.fill3DRect(model.x, model.y, model.w, model.h, true);
+		g.fill3DRect((int)(model.x), (int)(model.y), (int)(model.w), (int)(model.h), true);
 		
 	}
